refactor(LottoPurchase): Lotto 클래스를 구매와 당첨 확인으로 두 기능으로 분리.
